Implements a working (still minimal) video-player.
QMediaPlayer can’t play straight from QByteArray, as it needs a Qual
for the resource. Therefore I create a HTTPS server that serves the
content of the QByteArray to the player. The idea was to use
certificates to both server and client auth, but QMediaPlayer doesn’t
respect QSslConfiguration either. So now I serve the file using it’s
SHA3-512 as path, inside a TLS connection. Should be strong enough to
protect from loopback eavesdropping. On macOS, thanks to Apple’s
Security framework, an annoying password window appears when I try to
set the thrust of the CA cert, which apples obligates me to save in the
Keychain.